requirements.txt
# Core Dependencies
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging & Monitoring
tensorboard>=2.14.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
